fix(web): move blog frontmatter parse out of the Vite plugin
Types CI cannot import plugins/blog.ts. The homepage metadata schema
lives in lib so allowlist tests can cover it. Test fixtures take
partial homepage fields without requiring tags.
